Refactor internal methods into debugging utilities
[ghstack-poisoned]
