+++ torch/fx/graph.py
@@ -1,11 +1,11 @@
 from .node import Node, Argument, Target, map_arg
+from ._print_utils import _type_repr, _get_qualified_name
 
 from typing import Callable, Any, List, Dict, Optional, Tuple, Set
-import builtins
 import torch
-import types
 import keyword
 import re
+import builtins
 
 def _shadows_builtin_name(name: str) -> bool:
     return name in builtins.__dict__ or name in keyword.kwlist or name in {'inf', 'nan', 'NoneType'}
@@ -31,26 +31,6 @@ def _snake_case(s: str) -> str:
         prev_lower = c.islower()
     return ''.join(chars)
 
-def get_qualified_name(func: Callable[..., Any]) -> str:
-    # things like getattr just appear in builtins
-    if getattr(builtins, func.__name__, None) is func:
-        return func.__name__
-    name = func.__name__
-    module = _find_module_of_method(func)
-    module = module.replace('torch._ops', 'torch.ops')  # WAR for bug in how torch.ops assigns module
-    return f'{module}.{name}'
-
-# this is fixed on master, WAR for 1.5
-def _find_module_of_method(orig_method: Callable[..., Any]) -> str:
-    name = orig_method.__name__
-    module = orig_method.__module__
-    if module is not None:
-        return module
-    for guess in [torch, torch.nn.functional]:
-        if getattr(guess, name, None) is orig_method:
-            return guess.__name__
-    raise RuntimeError(f'cannot find module for {orig_method}')
-
 def _format_args(args: Tuple[Argument, ...], kwargs: Dict[str, Argument]) -> str:
     args_s = ', '.join(repr(a) for a in args)
     kwargs_s = ', '.join(f'{k} = {repr(v)}' for k, v in kwargs.items())
@@ -67,29 +47,6 @@ def _format_target(base: str, target: str) -> str:
         else:
             r = f'{r}.{e}'
     return r
-
-# Borrowed from CPython typing module
-# https://github.com/python/cpython/blob/f90dc36c15d7fee0efaf6d39e97be0bdf2683e93/Lib/typing.py#L156
-def _type_repr(obj):
-    """Return the repr() of an object, special-casing types (internal helper).
-    If obj is a type, we return a shorter version than the default
-    type.__repr__, based on the module and qualified name, which is
-    typically enough to uniquely identify a type.  For everything
-    else, we fall back on repr(obj).
-    """
-    # HACK: In Python 3.6, type aliases from ``typing`` are instances of ``type``, but in
-    # later Python versions, type aliases are not instances of ``type``!! We want
-    # all type aliases to fall through to ``repr``, so if we have a type that is
-    # in the module typing, don't go down this path.
-    if isinstance(obj, type) and obj.__module__ != 'typing':
-        if obj.__module__ == 'builtins':
-            return obj.__qualname__
-        return f'{obj.__module__}.{obj.__qualname__}'
-    if obj is ...:
-        return('...')
-    if isinstance(obj, types.FunctionType):
-        return obj.__name__
-    return repr(obj)
 
 class _InsertPoint:
     def __init__(self, graph, new_insert):
@@ -671,7 +628,7 @@ class Graph:
                     assert isinstance(node.args, tuple)
                     body.append(f'{node.name} = {magic_methods[node.target.__name__].format(*(repr(a) for a in node.args))}')
                     return
-                qualified_name = get_qualified_name(node.target)
+                qualified_name = _get_qualified_name(node.target)
                 register_modules_used(qualified_name)
                 if qualified_name == 'getattr' and \
                    isinstance(node.args, tuple) and \
@@ -733,66 +690,7 @@ def forward(self, {', '.join(free_vars)}){maybe_return_annotation[0]}:
         # over value
         maybe_return_typename : List[str] = ['']
 
-        def format_arg(arg) -> str:
-            if isinstance(arg, list):
-                items = ', '.join(format_arg(a) for a in arg)
-                return f'[{items}]'
-            elif isinstance(arg, tuple):
-                items = ', '.join(format_arg(a) for a in arg)
-                maybe_comma = ',' if len(arg) == 1 else ''
-                return f'({items}{maybe_comma})'
-            elif isinstance(arg, dict):
-                items_str = ', '.join(f'{k}: {format_arg(v)}' for k, v in arg.items())
-                return f'{{{items_str}}}'
-
-            if isinstance(arg, Node):
-                return '%' + str(arg)
-            else:
-                return str(arg)
-
-        def pretty_print_target(target):
-            """
-            Make target printouts more user-friendly.
-            1) builtins will be printed as `builtins.xyz`
-            2) operators will be printed as `operator.xyz`
-            3) other callables will be printed with qualfied name, e.g. torch.add
-            """
-            if isinstance(target, str):
-                return target
-            if hasattr(target, '__module__'):
-                if not hasattr(target, '__name__'):
-                    # Just to be defensive, if we don't have `__name__`, get the
-                    # qualname. Not sure if this happens for any members of `operator`
-                    # or `builtins`. This fallback path is not as good, since e.g.
-                    # things in `operator` have `_operator` as their __module__.
-                    return get_qualified_name(target)
-                if target.__module__ == 'builtins':
-                    return f'builtins.{target.__name__}'
-                elif target.__module__ == '_operator':
-                    return f'operator.{target.__name__}'
-            return get_qualified_name(target)
-
-        def format_node(n : Node) -> Optional[str]:
-            if n.op == 'placeholder':
-                assert isinstance(n.target, str)
-                arg_str = n.target
-                arg_str += arg_str + f': {_type_repr(n.type)}' if n.type is not None else ''
-                placeholder_names.append(arg_str)
-                return None
-            elif n.op == 'get_attr':
-                maybe_typename = f'{_type_repr(n.type)} ' if n.type is not None else ''
-                return f'%{n.name} : {maybe_typename}[#users={len(n.users)}] = self.{n.target}'
-            elif n.op == 'output':
-                if n.type is not None:
-                    maybe_return_typename[0] = f' -> {_type_repr(n.type)}'
-                return f'return {n.args[0]}'
-            else:
-                maybe_typename = f'{_type_repr(n.type)} ' if n.type is not None else ''
-                return f'%{n.name} : {maybe_typename}[#users={len(n.users)}] = {n.op}[target={pretty_print_target(n.target)}](' \
-                       f'args = {format_arg(n.args)}, kwargs = {format_arg(n.kwargs)})'
-
-
-        node_strs = [format_node(node) for node in self.nodes]
+        node_strs = [node.format_node(placeholder_names) for node in self.nodes]
         param_str = ', '.join(placeholder_names)
         s = f'graph({param_str}){maybe_return_typename[0]}:'
         for node_str in node_strs:
