+++ torch/fx/node.py
@@ -1,6 +1,7 @@
 # Nodes represent a definition of a value in our graph of operators.
 from typing import TYPE_CHECKING, Union, Callable, Any, Tuple, List, Optional, Dict
 from .immutable_collections import immutable_dict, immutable_list
+from ._print_utils import _type_repr, _get_qualified_name
 import torch
 
 if TYPE_CHECKING:
@@ -19,6 +20,23 @@ Argument = Optional[Union[
     'Node',
     BaseArgumentTypes
 ]]
+
+def _format_arg(arg) -> str:
+    if isinstance(arg, list):
+        items = ', '.join(_format_arg(a) for a in arg)
+        return f'[{items}]'
+    elif isinstance(arg, tuple):
+        items = ', '.join(_format_arg(a) for a in arg)
+        maybe_comma = ',' if len(arg) == 1 else ''
+        return f'({items}{maybe_comma})'
+    elif isinstance(arg, dict):
+        items_str = ', '.join(f'{k}: {_format_arg(v)}' for k, v in arg.items())
+        return f'{{{items_str}}}'
+
+    if isinstance(arg, Node):
+        return '%' + str(arg)
+    else:
+        return str(arg)
 
 class Node:
     """
@@ -77,7 +95,7 @@ class Node:
         #
         # For placeholder nodes, this value will be used to type-annotate the
         # generated function parameters.
-        # For the return ndoe, this value will be used to type-annotate the
+        # For the return node, this value will be used to type-annotate the
         # generated function return type. (Note this is a special case. ``return``
         # does not produce a value, it's more of a notation. Thus, this value
         # describes the type of args[0] in the ``return`` node.
@@ -219,6 +237,54 @@ class Node:
 
     def __repr__(self) -> str:
         return self.name
+
+    def _pretty_print_target(self, target):
+        """
+        Make target printouts more user-friendly.
+        1) builtins will be printed as `builtins.xyz`
+        2) operators will be printed as `operator.xyz`
+        3) other callables will be printed with qualfied name, e.g. torch.add
+        """
+        if isinstance(target, str):
+            return target
+        if hasattr(target, '__module__'):
+            if not hasattr(target, '__name__'):
+                # Just to be defensive, if we don't have `__name__`, get the
+                # qualname. Not sure if this happens for any members of `operator`
+                # or `builtins`. This fallback path is not as good, since e.g.
+                # things in `operator` have `_operator` as their __module__.
+                return _get_qualified_name(target)
+            if target.__module__ == 'builtins':
+                return f'builtins.{target.__name__}'
+            elif target.__module__ == '_operator':
+                return f'operator.{target.__name__}'
+        return _get_qualified_name(target)
+
+    def format_node(self, placeholder_names=None, maybe_return_typename=None) -> Optional[str]:
+        """
+        Return a descriptive string representation of ``self``
+        """
+        if self.op == 'placeholder':
+            assert isinstance(self.target, str)
+            arg_str = self.target
+            arg_str += arg_str + f': {_type_repr(self.type)}' if self.type else ''
+            if placeholder_names:
+                placeholder_names.append(arg_str)
+                return None
+            maybe_typename = f'{_type_repr(self.type)} ' if self.type else ''
+            default_val = '(default=' + str(self.args[0]) + ')' if self.args else ''
+            return f'%{self.name} : {maybe_typename}[#users={len(self.users)}] = {self.op}[target={self.target}]{default_val}'
+        elif self.op == 'get_attr':
+            maybe_typename = f'{_type_repr(self.type)} ' if self.type is not None else ''
+            return f'%{self.name} : {maybe_typename}[#users={len(self.users)}] = self.{self.target}'
+        elif self.op == 'output':
+            if self.type and maybe_return_typename:
+                maybe_return_typename[0] = f' -> {_type_repr(self.type)}'
+            return f'return {self.args[0]}'
+        else:
+            maybe_typename = f'{_type_repr(self.type)} ' if self.type is not None else ''
+            return f'%{self.name} : {maybe_typename}[#users={len(self.users)}] = {self.op}[target={self._pretty_print_target(self.target)}](' \
+                   f'args = {_format_arg(self.args)}, kwargs = {_format_arg(self.kwargs)})'
 
     def replace_all_uses_with(self, replace_with : 'Node') -> List['Node']:
         """
